Remove/restore unworking implementation attempts
Remove implementation attempts that didn't work, restoring those files.
