+++ satpy/readers/xmlformat.py
@@ -19,26 +19,25 @@
 
 from __future__ import annotations
 
-from xml.etree.ElementTree import ElementTree  # nosec
+from xml.etree.ElementTree import ElementTree
 
 import numpy as np
+
+VARIABLES: dict[str, str] = {}
 
 TYPEC = {"boolean": ">i1",
          "integer2": ">i2",
          "integer4": ">i4",
-         "vinteger4": ">i4",
-         "uinteger1": ">u1",
          "uinteger2": ">u2",
-         "vuinteger2": ">u2",
-         "uinteger4": ">u4",
-         "vuinteger4": ">u4"}
+         "uinteger4": ">u4", }
 
 
-def process_delimiter(elt, variables, ascii=False):
+def process_delimiter(elt, ascii=False):
     """Process a 'delimiter' tag."""
+    del elt, ascii
 
 
-def process_field(elt, variables, ascii=False):
+def process_field(elt, ascii=False):
     """Process a 'field' tag."""
     # NOTE: if there is a variable defined in this field and it is different
     # from the default, we could change the value and restart.
@@ -66,26 +65,28 @@ def process_field(elt, variables, ascii=False):
     return ((elt.get("name"), current_type, scale))
 
 
-def process_array(elt, variables, ascii=False):
+def process_array(elt, ascii=False):
     """Process an 'array' tag."""
+    del ascii
     chld = list(elt)
     if len(chld) > 1:
         raise ValueError()
     chld = chld[0]
     try:
-        name, current_type, scale = CASES[chld.tag](chld, variables, ascii)
+        name, current_type, scale = CASES[chld.tag](chld)
         size = None
     except ValueError:
-        name, current_type, size, scale = CASES[chld.tag](chld, variables, ascii)
+        name, current_type, size, scale = CASES[chld.tag](chld)
+    del name
     myname = elt.get("name") or elt.get("label")
     if elt.get("length").startswith("$"):
-        dimname = elt.get("length")[1:]
-        length = int(variables[dimname])
+        length = int(VARIABLES[elt.get("length")[1:]])
     else:
         length = int(elt.get("length"))
     if size is not None:
         return (myname, current_type, (length, ) + size, scale)
-    return (myname, current_type, (length, ), scale)
+    else:
+        return (myname, current_type, (length, ), scale)
 
 
 CASES = {"delimiter": process_delimiter,
@@ -138,6 +139,36 @@ def to_scales(val):
     return scales
 
 
+def parse_format(xml_file):
+    """Parse the xml file to create types, scaling factor types, and scales."""
+    tree = ElementTree()
+    tree.parse(xml_file)
+    for param in tree.find("parameters"):
+        VARIABLES[param.get("name")] = param.get("value")
+
+    types_scales = {}
+
+    for prod in tree.find("product"):
+        ascii = (prod.tag in ["mphr", "sphr"])
+        res = []
+        for i in prod:
+            lres = CASES[i.tag](i, ascii)
+            if lres is not None:
+                res.append(lres)
+        types_scales[(prod.tag, int(prod.get("subclass")))] = res
+
+    types = {}
+    stypes = {}
+    scales = {}
+
+    for key, val in types_scales.items():
+        types[key] = to_dtype(val)
+        stypes[key] = to_scaled_dtype(val)
+        scales[key] = to_scales(val)
+
+    return types, stypes, scales
+
+
 def _apply_scales(array, scales, dtype):
     """Apply scales to the array."""
     new_array = np.empty(array.shape, dtype)
@@ -152,14 +183,12 @@ def _apply_scales(array, scales, dtype):
     return new_array
 
 
-class XMLFormat:
+class XMLFormat(object):
     """XMLFormat object."""
 
     def __init__(self, filename):
         """Init the format reader."""
-        self.dims = {}
-        self.variables = {}
-        self.types, self.stypes, self.scales = self.parse_format(filename)
+        self.types, self.stypes, self.scales = parse_format(filename)
 
         self.translator = {}
 
@@ -174,31 +203,6 @@ class XMLFormat:
         """Apply scales to *array*."""
         return _apply_scales(array, *self.translator[array.dtype])
 
-    def parse_format(self, xml_file):
-        """Parse the xml file to create types, scaling factor types, and scales."""
-        tree = ElementTree()
-        tree.parse(xml_file)
-        for param in tree.find("parameters"):
-            self.variables[param.get("name")] = param.get("value")
 
-        types_scales = {}
-
-        for prod in tree.find("product"):
-            ascii = (prod.tag in ["mphr", "sphr"])
-            res = []
-            for i in prod:
-                lres = CASES[i.tag](i, self.variables, ascii)
-                if lres is not None:
-                    res.append(lres)
-            types_scales[(prod.tag, int(prod.get("subclass")))] = res
-
-        types = {}
-        stypes = {}
-        scales = {}
-
-        for key, val in types_scales.items():
-            types[key] = to_dtype(val)
-            stypes[key] = to_scaled_dtype(val)
-            scales[key] = to_scales(val)
-
-        return types, stypes, scales
+if __name__ == '__main__':
+    pass
